tui/requirements.txt
textual>=0.47.0
psycopg2-binary>=2.9.0
plotext>=5.3.0
